requirements.txt
sys
moviepy
pillow
openai
elevenlabs
opencv-python
pydub
numpy
#-e .